Create a quick path past Homebrew Installation
For users who know how to open a terminal, they may know already that
they have Homebrew installed.  Allow them to quickly state that and move
on.  If they needed help opening the terminal, we continue to help them
asses whether Homebrew is installed.

When asking if installing Homebrew was successful, skip verification on
"yes", but add "I don't know" which checks.

Co-authored-by: Lisa Ugray <lisa.ugray@gmail.com>