Fix reasonable solutions
The first reasonable solution is missing parentheses, which means that when the left hand of the logical AND short-cirtcuits, the right hand of the logical OR is always executed. The left hand short-circuits if a year is not divisible by 4, so the right hand of the logical OR will ALWAYS be false (as it can't be divisible by 400 then).