Remove explicit module notion and make it a first class citizen
At this point in the track it's fine to use ANYTHING, but the notion that "class" is somehow more correct than module is off-putting.

A class should only be used if its functionality requires instantation. Utility functions and the sorts should either be on a named constant which is an Object, or a module so it can be included/extended and shared in other modules and classes.