Add more reasonable solutions including string templates.
The point of this exercise is to prepare a student for the track, not force one usage over another.

String interpolation is simpler, but template strings are much more maintainable when used with named arguments as it can be subtituted easily for internationalisation and/or extended.

This adds all the variations as reasonable so mentors no longer penalise students for using it.