add blurb to keep [ and ] chars in json file
There have been a few occasions where the first mentor to a JSON file has removed the square brackets that are intended to encapsulate the mentors that may be added to a particular JSON file.  The blurb I have added is intended to indicate that the square brackets should not be removed.
add JSON array clarification text.