Update code_of_conduct.md
Modify a couple of lines in the 'We condemn' section, in order to make the style consistent (those lines didn't follow the "we condemn..." style).

Compress two separate 'disrespect' lines into one, as it seems a bit pointless to have the two of them apart.

Capitalise the start of each entry in the 'The Simple Version' and 'The Details' lists, in order to make them consistent with the other lists on the page.

Write out Code of Conduct in full the first time it is used on the page with the contraction in parentheses following it.  I know this is arguably a bit pointless when the page is titled Code of Conduct, but it's considered best-practice style in pretty much every field and venue I have written for in the past.