Make links to act like links on Mentoring FAQ
While it looks good here on Github, on the [website page](https://exercism.io/mentoring-faqs) URLs that aren't formatted as such don't become hyperlinks that you can click. This gives me very unpleasant experience as a reader, I need to copy it and paste in the new tab URL bar. My "fix" is to format URLs in Markdown in such a way that static pages generator would recognize the links.