javascript/leap: round 2 note improvements (#780)
* Add summary of goals

Similar to the other JavaScript mentor notes, add a lead that links to reference material about the goals of the exercise.

* Fix reasonable solutions

The first reasonable solution is missing parentheses, which means that when the left hand of the logical AND short-cirtcuits, the right hand of the logical OR is always executed. The left hand short-circuits if a year is not divisible by 4, so the right hand of the logical OR will ALWAYS be false (as it can't be divisible by 400 then).

* Add note about types of solutions

Most common thing is not using logical operators, or having a clever solution. Instruct the mentor to fix that first. A lot of the times you don't need to mentor anything else and it will fix it by itself.

* Add more guidance when to give a suggestion

Have the example in the common suggestions, so a mentor can quickly scan these notes and not need to apply each line by itself.

* Add suggestion about order of evaluation

Operator precedence and short-circuiting is important in this exercise. Not just order of operations.

* Add reference links to all the talking points, where possible

* Add solution using a function generator

* Add implicit thruthiness example

* Change functional solution to not use advanced topics such as currying

* javascript/leap: round 2 notes

Improves the notes with:
- templates to copy and paste
- instructions on when to approve and when to only suggest
- more suggestions around short-circtuiting, ordering and parentheses uses

All these changes are based on 200+ mentoring sessions.

* Make the templates copyable

The quoted text looks much beter, but doesn't allow for copy and paste.

* Fix typo

* Implement requested changes