Improve external reference
This updates the link to the LaunchSchool page to point directly to the section that discusses the advantages of referencing a getter method internally in a class as opposed to referencing the instance variable directly. Previously, it was pointing to a section higher up on the page, and I had to read for quite a while before I got to the relevant section.

It also adds some additional explanation about what the link explains. The previous "This link explains:" with no additional detail seemed to vague to me.