each_char.with_index vs chars.each_with_index
Can we all agree this is better as it avoid the need to create the extra array?  This brings solution 3 more in line with solution 2.  In my quick performance tests with small strings solution 3 (with each_char) seems to beat them all.

Or should we add this as a 4th strategy?  It's obviously the correct version of the 3rd strategy though IMHO.