Ruby/Luhn New notes
New notes!

I'd appreciate if we use this PR for correcting errors only. There's some urgency to merge, because there are several Luhns in the queue waiting for mentoring, and thus a great test case for the Notes. 

I opened a separate issue at the same time, where we can discuss the next iteration of the Luhn Notes.