Made the code simpler and more concise.
It's encouraged to shadow a variable name, if the name is appropriate and the old value isn't needed anymore.

`match` shouldn't be used for single booleans, `if`/`else` is equally readable, but shorter and requires fewer indents and braces.

It's recommended to use functions paths instead of closures where possible.

Rust can infer types, so it's OK to omit the type, if it's obvious from the variable name. Variables starting with `is_` or `contains_` are usually booleans.