Clarify mutation is allowed in high-scores
Adding a test to check against mutating the list has been suggested and rejected because this is an introductory core exercise and immutability has not yet been introduced. However, all of the reasonable examples provided do not mutate the list so it is not clear to new mentors whether preserving the original list is required when deciding if they should accept a solution.