python/raindrops: Simplify notes for Track Anatomy re-order (#1206)
* python raindrops: simplify mentoring.md

Simplifies the mentoring.md to address new position in the track
core. Removes several side concerns that assume more complexity than
is required.

* python raindrops: add ternary if and fix wording

Adds an example for the ternary if and fixes and clarifies some wording.

* python raindrops: readability fixes

Fixes miscellanous issues not visible until Markdown was rendered.

* raindrops: add type annotation note

Updates the Talking Points to add a note about type annotations.