Fixing module name for consistency to problem file (#1388)
There was a recent change in the elixir track that proposed to change the `Words` module to `WordCount` for better congruity with the problem name, slug.  This is because some vi and emacs setups use the filename to determine which file(s) to use for testing rather than relying on `mix test` to find them all.