Rust(Clock): Recommend rem/div_euclid more
As Rust's current version is v1.42, comments about Euclidean division being
unstable are outdated. While using % or / is acceptable, rem/div_euclid are
preferable for this exercise. This patch adjusts suggestions to reflect that, and
removes suggestions that might be redundant.

It also continues tweaking line width of suggestions re: helper functions for
the convenience of the mentor reading them.